+++ addOns/callhome/src/main/java/org/zaproxy/addon/callhome/ExtensionCallHome.java
@@ -332,6 +332,7 @@ public class ExtensionCallHome extends ExtensionAdaptor
                     || key.startsWith("stats.sequence.")
                     || key.startsWith("stats.spider.")
                     || key.startsWith("stats.spiderAjax.")
+                    || key.startsWith("stats.sse.")
                     || key.startsWith("stats.tech.")
                     || key.startsWith("stats.ui.")
                     || key.startsWith("stats.websockets.")